+++ lib/src/messages/languages/ar_msg.dart
@@ -1,10 +1,11 @@
+import '../future_time_messages.dart';
 import '../../utils/utility.dart';
 import '../messages.dart';
 
 /// The [ArabicMessages] class provides Arabic language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in Arabic.
-class ArabicMessages extends Messages {
+class ArabicMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => 'قبل';