+++ lib/src/messages/languages/en_msg.dart
@@ -1,9 +1,10 @@
+import '../future_time_messages.dart';
 import '../messages.dart';
 
 /// The [EnglishMessages] class provides English language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in English.
-class EnglishMessages implements Messages {
+class EnglishMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => '';