+++ lib/src/messages/languages/es_msg.dart
@@ -1,9 +1,10 @@
+import '../future_time_messages.dart';
 import '../messages.dart';
 
 /// The [EspanaMessages] class provides Spanish language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in Spanish.
-class EspanaMessages implements Messages {
+class EspanaMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => 'hace';