+++ lib/src/messages/languages/it_msg.dart
@@ -1,9 +1,10 @@
+import '../future_time_messages.dart';
 import '../messages.dart';
 
 /// The [ItalianMessages] class provides Italian language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in Italian.
-class ItalianMessages implements Messages {
+class ItalianMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => '';