+++ lib/src/messages/languages/ko_msg.dart
@@ -1,9 +1,10 @@
+import '../future_time_messages.dart';
 import '../messages.dart';
 
 /// The [KoreanMessages] class provides Korean language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in Korean.
-class KoreanMessages implements Messages {
+class KoreanMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => '';