+++ lib/src/messages/languages/vi_msg.dart
@@ -1,9 +1,10 @@
+import '../future_time_messages.dart';
 import '../messages.dart';
 
 /// The [VietnameseMessages] class provides Vietnamese language-specific
 /// implementations of the [Messages] interface. This class is used to format
 /// the "time ago" strings in Vietnamese.
-class VietnameseMessages implements Messages {
+class VietnameseMessages with FutureTimeMessages implements Messages {
   /// Prefix added before the time message.
   @override
   String prefixAgo() => '';